Refactor package in ControllerTests
