+++ src/test/java/pro/sky/animal_shelter_telegram_bot/controller/PetControllerTests/PetControllerTest.java
@@ -1,38 +1,39 @@
-package pro.sky.animal_shelter_telegram_bot.controller;
+package pro.sky.animal_shelter_telegram_bot.controller.PetControllerTests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import pro.sky.animal_shelter_telegram_bot.controller.PetController;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static pro.sky.animal_shelter_telegram_bot.controller.ConstantsForControllerTests.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class PetOwnerControllerTest {
+class PetControllerTest {
 
     @LocalServerPort
     private int port;
 
     @Autowired
-    private PetOwnerController petOwnerController;
+    private PetController petController;
 
     @Autowired
     private TestRestTemplate restTemplate;
 
     @Test
     public void contextLoads(){
-        assertThat(petOwnerController).isNotNull();
+        assertThat(petController).isNotNull();
     }
 
     @Test
     public void testHelloMessage(){
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_URL + "/", String.class))
                 .isNotEmpty();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_URL + "/", String.class))
                 .isNotNull();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_OWNER_URL + "/", String.class))
-                .isEqualTo(HELLO_MESSAGE_PET_OWNER);
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + PET_URL + "/", String.class))
+                .isEqualTo(HELLO_MESSAGE_PET_CONTROLLER);
     }
 }
