+++ src/test/java/pro/sky/animal_shelter_telegram_bot/controller/ReportControllerTests/ReportControllerTest.java
@@ -1,38 +1,39 @@
-package pro.sky.animal_shelter_telegram_bot.controller;
+package pro.sky.animal_shelter_telegram_bot.controller.ReportControllerTests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import pro.sky.animal_shelter_telegram_bot.controller.ReportController;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static pro.sky.animal_shelter_telegram_bot.controller.ConstantsForControllerTests.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class VolunteerControllerTest {
+class ReportControllerTest {
 
     @LocalServerPort
     private int port;
 
     @Autowired
-    private VolunteerController volunteerController;
+    private ReportController reportController;
 
     @Autowired
     private TestRestTemplate restTemplate;
 
     @Test
     public void contextLoads(){
-        assertThat(volunteerController).isNotNull();
+        assertThat(reportController).isNotNull();
     }
 
     @Test
     public void testHelloMessage(){
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + VOLUNTEER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + REPORT_URL + "/", String.class))
                 .isNotEmpty();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + VOLUNTEER_URL + "/", String.class))
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + REPORT_URL + "/", String.class))
                 .isNotNull();
-        assertThat(this.restTemplate.getForObject(URL + port + "/" + VOLUNTEER_URL + "/", String.class))
-                .isEqualTo(HELLO_MESSAGE_VOLUNTEER_CONTROLLER);
+        assertThat(this.restTemplate.getForObject(URL + port + "/" + REPORT_URL + "/", String.class))
+                .isEqualTo(HELLO_MESSAGE_REPORT_CONTROLLER);
     }
 }
